+++ src/disseminate/tags/toc.py
@@ -103,12 +103,13 @@ class Toc(Tag):
         assert self.context.is_valid('label_manager')
 
         # If 'all' is specified in the toc_kind, then all documents should be
-        # selected. This is done by having a context of None with the
+        # selected. This is done by having a doc_id of None with the
         # 'get_labels' method of the label manager. If 'all' is not
-        # specified, then use this document's context. This will return labels
+        # specified, then use this document's doc_id. This will return labels
         # only for this document and its context from the 'get_labels' method
         # of the label manager.
-        context = self.context if 'all' not in self.toc_kind else None
+        context = self.context
+        doc_id = context.get('doc_id') if 'all' not in self.toc_kind else None
 
         # Create a default function for order labels. This may be overwritten
         # below, depending on the type of TOC.
@@ -124,8 +125,8 @@ class Toc(Tag):
 
         if 'heading' in self.toc_kind or 'headings' in self.toc_kind:
             # Retrieve heading labels, either for this document or all documents
-            # (depending on the value of context)
-            labels = label_manager.get_labels(context=context,
+            # (depending on the value of doc_id)
+            labels = label_manager.get_labels(doc_id=doc_id,
                                               kinds='heading')
 
             last_heading_level = None
@@ -153,19 +154,17 @@ class Toc(Tag):
             return labels, order_function, 'heading'
 
         if 'document' in self.toc_kind or 'documents' in self.toc_kind:
-            # Get the doc_id for the current document
-            doc_id = self.context['doc_id']
-
             # Retrieve the labels for the documents and the headings. Either
             # for this document or all documents (depending on the value of
-            # context)
-            document_labels = label_manager.get_labels(context=context,
+            # doc_id)
+            document_labels = label_manager.get_labels(doc_id=doc_id,
                                                        kinds='document')
-            heading_labels = label_manager.get_labels(context=context,
+            heading_labels = label_manager.get_labels(doc_id=doc_id,
                                                       kinds='heading')
 
             # Reorganize the document and heading labels such that the headings
             # are between documents
+            current_doc_id = self.context['doc_id']
             merged_labels = []
             for document_label in document_labels:
                 merged_labels.append(document_label)
@@ -175,7 +174,7 @@ class Toc(Tag):
                 # current document
                 if ('expanded' in self.toc_kind or
                    ('abbreviated' in self.toc_kind and
-                    document_label.doc_id == doc_id)):
+                    document_label.doc_id == current_doc_id)):
 
                     merged_labels += [l for l in heading_labels
                                       if l.doc_id == document_label.doc_id]
@@ -275,7 +274,8 @@ class Toc(Tag):
         self.ref_tags = tags
         self._mtime = latest_mtime
 
-    def html_fmt(self, content=None, listtype='ul', elements=None, level=1):
+    def html_fmt(self, content=None, listtype='ul', elements=None,
+                 toc_level=1, level=1):
         """Convert the tag to an html listing.
 
         .. note:: The 'document' toc is special since it uses the documents
@@ -291,7 +291,9 @@ class Toc(Tag):
         listtype : str, optional
             The type of list to render (ol, ul).
         elements : str, optional
-            The the reference tags.
+            The reference tags.
+        toc_level : int, optional
+            The level of the toc.
         level : int, optional
             The level of the tag.
 
@@ -327,7 +329,8 @@ class Toc(Tag):
                 # The element is a list of tags. Process this list as a group.
                 returned_elements.append(self.html_fmt(content=None,
                                                        elements=e,
-                                                       level=level + 1))
+                                                       level=level + 1,
+                                                       toc_level=toc_level + 1))
             else:
                 # Otherwise it's a ref tag, get its html and wrap it in a list
                 # item
@@ -338,7 +341,7 @@ class Toc(Tag):
 
         # Create list tag
         list_tag = html_tag(listtype,
-                            attributes='class=toc-level-{}'.format(level),
+                            attributes='class=toc-level-{}'.format(toc_level),
                             formatted_content=returned_elements,
                             level=level)
         return list_tag